requirements.txt
streamlit>=1.28.0
requests>=2.31.0
Pillow>=10.0.0
pandas>=2.0.0
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0
wordcloud>=1.9.0
